requirements.txt
# Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# AI - HuggingFace
requests==2.31.0
huggingface-hub==0.20.0

# Video Processing
moviepy==1.0.3
openai-whisper
librosa==0.10.1
soundfile==0.12.1
noisereduce==3.0.0

# NumPy - Fixed for Python 3.12
numpy>=1.26.0

# OpenCV - Compatible version
opencv-python==4.10.0.84

# Pillow
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0

# Frontend
streamlit==1.31.0

# Additional dependencies that might be needed
decorator==4.4.2
imageio==2.31.1
imageio-ffmpeg==0.4.9
proglog==0.1.10
tqdm==4.66.1